simple_machine_learning_deployments/requirements.txt
streamlit==1.24.1
scikit-learn==1.30.0
catboost==1.2.2
xgboost==2.0.2
pandas==2.1.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
python-environ==0.4.54